Fixing DRV8311 recursive design flaw, and pin tie issue
Since I'm using 3x pwm mode for the driver, I should have tied the unused lowside pins to 3.3v, and not left them floating, according to the datasheet. This is fixed now

The driver chip has an enable pin with an internal weak pulldown. I put a pull-up resistor on the outside, to always keep the chip on. Guess what that is connected to though? The regulator output from the chip, which is disabled! I fixed this in the design by have a voltage divider coming from motor power to enable the pin.