Working Magnetic sensor angle detection
It seems like the field strength value is really only an indicator if it is the correct distance. It is a 1 when to far, 0 when just right, and 2 when too close. I think I have to modify the source driver code, because though the sensor models are similar (MA730, mine: MA735), they have slightly different registers.