Finally Fixed motor sensor alignment issue
I had forgotten about Ohms law in setting the voltage sensor alignment, which lead to a current beyond my systems capacity. The result was probably capped waveforms and the motor not moving. 25 amps of current if the voltage is set to 5 on my motor! I now set the the voltage sensor align to 0.5, which gives about 2.5 amps current during alignment.