Testing STM32F405 to improve speed
I am using the Adafruit Feather board for this, wired up with a bunch of manually soldered wires to the test pads on the motor controller. My board is programmed to do nothing, so the Feather can control the driver instead. Its messy, but I REALLY want to get over 1000 radians a second closed loop. The feather is clocked at 168mhz.